+++ src/components/App.jsx
@@ -20,10 +20,10 @@ export default function App() {
     <div className="app">
       <Button {...buttonProps} clickHandler={editHandler} />
       <Header isEditing={isEditing} />
-      <Contact isEditing={isEditing} />
+      {/* <Contact isEditing={isEditing} /> */}
       <Education isEditing={isEditing} />
-      <Skills isEditing={isEditing} />
-      <Work isEditing={isEditing} />
+      {/* <Skills isEditing={isEditing} /> */}
+      {/* <Work isEditing={isEditing} /> */}
       {/* <References isEditing={isEditing} /> */}
     </div>
   );